__requirements.txt
-e .
attrs==20.2.0
bidict==0.21.2
cachetools==4.1.1
certifi==2020.6.20
cftime==1.2.1
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
colorama==0.4.3
configparser==5.0.0
connexion==2.7.0
cycler==0.10.0
decorator==4.4.2
flask-cors==3.0.9
flask-socketio==4.3.1
flask==1.1.2
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-secret-manager==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0; python_version < '3.8'
inflection==0.5.1
itsdangerous==1.1.0
jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
kiwisolver==1.2.0
lxml==4.5.2
mako==1.1.3
markdown==3.2.2
markupsafe==1.1.1
matplotlib==3.3.1
minio==6.0.0
neo4j-driver==4.1.1
netcdf4==1.5.4
numpy==1.19.2
openapi-spec-validator==0.2.9
pandas==1.1.2
passlib==1.7.2
pdoc3==0.9.1
pillow==7.2.0
prance==0.19.0
protobuf==3.13.0
psycopg2-binary==2.8.6
py==1.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.17.2
pyshp==2.1.2
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
pyyaml==5.3.1
redis==3.5.3
requests==2.24.0
retry==0.9.2
rsa==4.6; python_version >= '3.5'
scikit-learn==0.23.2
scipy==1.5.2
semver==2.10.2
six==1.15.0
sqlalchemy==1.3.19
threadpoolctl==2.1.0
urllib3==1.24.3
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.35.1
zipp==3.1.0
